fix(launcher): shrink launcher-generator under its file-size ceiling (PAN-3300)
`main` was RED: the file-size guard failed because PAN-1837 landed
src/lib/launcher-generator.ts at 1018 lines against the 1012-line audited
exception it wrote for itself in scripts/file-size-allowlist.txt.

Shrink rather than raise the ceiling, per the guard's shrink-only contract:

- extract the Codex command shapes (buildCodexCommand and its token
  computation) into src/lib/launcher-codex-command.ts. It stays a leaf — the
  generator passes the LauncherConfig subset those shapes read plus a
  supervisor-wrap callback, so nothing imports back into the generator and no
  new import cycle appears.
- extract the shared shellQuote helper into src/lib/shell-quote.ts so both
  modules use one implementation instead of a copy.

launcher-generator.ts is now 892 lines; the allowlist entry stays at 1012.
